Server and Client threads work BUT problem with sending messages
The problem is, if TCP_Client has to become a thread, starting a thread isn't enough. Since the class TCP_Client has two functions (connect & send) which must have a common variable (the outputstream), there needs to be a way to be able to start a certain portion of the thread (like two runnables in a single thread?)
The connect and send functions cannot be seperated because the variable output stream has to be used by the send function after it has been initialized by the connect function.
The connect function is going to be executed just one (when hitting the CHAT button) whereas the send button is going to be executed each time the user hits SEND.
Summary:
-Two functions that need to be in the same class (to have a common variable) have to be executed separately but that is not possible using threads.