Added unique username function BUT problem with connecting 2 clients at the same time. Will continue debugging tomorrow.
While is a blocking process. So, instead of waiting for the client to type a unique username, we can use a thread to do just that: it waits until the username is valid to then add the client thread in the threads list. But, in doing so, the program malfunctions.
Needs more debugging.
Will continue debugging tomorrow.